src/requirements.txt
Django
django-cors-headers
djangorestframework
djongo
dnspython
pymongo